FX more closely match serato behavior
First 2 knobs on each deck, so 4 total, used to control each FX unit.
The routing matrix now like serato's, where all fx on a particular side
are applied to a deck if that deck's button is lit up.

This is fairly simple, though not as simple for the user as before. The
main motivation for this change was to pick a design that mixxx could
more easily support.

Previous commit emulated per-logical-deck FX knobs, which required doing
some stuff like cycling an effectunit an assumed-correct number of
times. The effects would glitch out sometimes, in that I would lose
sound from one deck and would have to futz with toggling effects and the
built-in filter (QuickEffect or something) to get it to come back (took
some time). And if a user cycled any of the units themselves, it would
break in rather boring ways.