Generate sp1.js
Rather than having to develop one monolithic file, we can
separate them into more manageable pieces, and using a primitive
inclusion system, combine them into the final result.

I have decided to include sp1.js in the git repo so that others who are
not interested in hacking on this script can just git-clone (or download
directly from github), copy the xml+sp1.js files into their controllers
directory, and be on their way.

I am aware that there are more sophisticated tools to combine js files
for distribution. But the needs of this project are so trivial, I'd
rather write a couple lines of awk and avoid dependency. Keep it simple.