Midi message revamp
My original methodology for gathering midi messages (namely, running
amidi and pressing the controls I was interested in) was done under the
assumption that the logical deck features of the sp1 were an illusion
provided by the software; so if you light up the deck 3 button, all the
controls on the left will send the same midi messages as deck 1.

Once it became clear that this is not the case (i.e., the sp1 itself
keeps track of which deck you're using and outputs midi messages
accordingly), it was clear another methodology for gathering midi
messages would be necessary. Pioneer publishes a pdf with the midi
mappings, but this format is unsuitable for consumption by a script, and
I wanted to get all the midi mappings into the program once and for all.

I wrote a script that guides you through the process of 'jiggering'
every control in every state (all decks, all pad modes, with and without
shift). that script was transformed into the new sp1-midiMap.js.

Since my original model of the sp1's midi messages was wrong, the
physKey conventions I was using had to be changed, and certain controls
that were handled by the decks before (e.g., fxknob 3 for the deck's
filter) had to be handled by the component that actually receives the
midi messages (so the fx component now receives fxknob 3, and forwards
the midi signal to whatever deck is currently selected on the left).

While I was at it, I also made an effort to convert to using
midiValueHandler() wherever possible. Testing with decks 3 and 4 has
gone smoothly.

So this commit makes deck 3/4 work, along with fx assignment.

The script currently requires you to press shift+deck3 or shift+deck4 to
go back to deck1/2. I may want to switch this to a toggle behavior, but
I'll have to play around a bit more before I commit to that.