Slicer support
Slicer takes a section of the track, divides it into
equal slices, and lets you hold any of the 8 slicer pads
to loop that slice (in slip mode).

- If no slicer position is set, press any of the slicer
pads to select the slice size (same sizes as hotcue). That
pad will switch its LED on to let you know there is an
active slicer position.
- If a slicer position _is_ set, press any of the slicer
pads to loop that pad's slice.
- shift+pad changes the slice size (all slices are adjusted). This
will NOT change the slicer position (unless there is no position
yet, in which case it will be set)
- shift+slicer (aka manualloop) clears the slicer position; you
can now press a pad again to reset the slicer position and slice
size.