Slicer v2
The previous slicer was a bit more static: it would pick a certain
track position and slice that position. While this vaguely approximates
the serato slicer, a key feature of the serato slicer is that it has a
"rolling" slicer position. This commit attempts to emulate that
behavior.

There are some finnicky behaviors that make this hard to do generally.
While slices _generally_ work, there are some caveats:

- a 'reference downbeat' is required. The slice section is computed
relative to this reference downbeat. We use hotcue 36 for the reference
downbeat. If there is no hotcue 36, we will use the first hotcue that is
set.
- use shift+slicer to explicitly set a reference downbeat.
- the slicer basically requires quantization to be on
- To smoothly change loops, it seems best to toggle on a beatloop at the
current play position, then move the loop to the desired position. This
requires doing a lot of math to figure out where the loop is now, and
then move it to the desired location.
- It seems that this math can go wrong if you switch pads fast enough.